dbms_agent_python/dbms_agent/semantic_model.py
# dbms_agent/semantic_model.py

"""
Semantic Model Analysis: DDL/DML analysis, schema embedding, and vector store integration.

Note: Azure Cosmos DB is the selected and only backend for embedding storage (Azure AI Search option deprecated).
"""



import os
import requests
from azure.cosmos import CosmosClient, PartitionKey

class SemanticModelAnalyzer:
    def __init__(self, db_connection):
        self.db_connection = db_connection
        # Cosmos DB config from environment variables (lazy init)
        self.cosmos_endpoint = os.getenv('COSMOS_ENDPOINT')
        self.cosmos_key = os.getenv('COSMOS_KEY')
        self.cosmos_db = os.getenv('COSMOS_DATABASE', 'dbms_agent')
        self.cosmos_container = os.getenv('COSMOS_CONTAINER', 'schema_vectors')
        self._cosmos_initialized = False
        self.client = None
        self.database = None
        self.container = None

    def _ensure_cosmos(self):
        if self._cosmos_initialized:
            return
        if not self.cosmos_endpoint or not self.cosmos_key:
            raise ValueError("COSMOS_ENDPOINT and COSMOS_KEY must be set for embedding storage.")
        self.client = CosmosClient(self.cosmos_endpoint, self.cosmos_key)
        self.database = self.client.create_database_if_not_exists(self.cosmos_db)
        self.container = self.database.create_container_if_not_exists(
            id=self.cosmos_container,
            partition_key=PartitionKey(path="/table_name")
        )
        self._cosmos_initialized = True

    def analyze_schema(self):
        """Analyze DDL and DML, generate schema representation."""
        # Placeholder: fetch schema from db_connection
        # Return a list of table/column dicts
        return []

    def embed_schema(self, schema=None):
        """Generate and store schema embeddings in Cosmos DB as a vector store."""
        if schema is None:
            schema = self.analyze_schema()
        # Placeholder: generate embeddings for each table/column
        self._ensure_cosmos()
        for table in schema:
            embedding = self._generate_embedding(table)
            item = {
                'id': table['name'],
                'table_name': table['name'],
                'embedding': embedding,
                'metadata': table
            }
            self.container.upsert_item(item)

    def _generate_embedding(self, table):
        """
        Generate embedding for a table using Azure OpenAI text-embedding-3-small model.
        """
        endpoint = os.getenv('AZURE_OPENAI_ENDPOINT')
        api_key = os.getenv('AZURE_OPENAI_API_KEY')
        deployment = os.getenv('AZURE_OPENAI_EMBEDDING_DEPLOYMENT', 'text-embedding-3-small')
        if not endpoint or not api_key:
            raise ValueError("Azure OpenAI endpoint and API key must be set in environment variables.")
        headers = {
            "Content-Type": "application/json",
            "api-key": api_key
        }
        url = f"{endpoint}/openai/deployments/{deployment}/embeddings?api-version=2024-02-15-preview"
        # Use table name and columns as embedding input
        text = f"Table: {table['name']}, Columns: {', '.join(table.get('columns', []))}"
        data = {"input": text}
        response = requests.post(url, headers=headers, json=data)
        response.raise_for_status()
        embedding = response.json()["data"][0]["embedding"]
        return embedding
