dbms_agent_python/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
azure-identity>=1.15.0
azure-cosmos>=4.6.0
pyodbc>=5.0.1
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0
